LRCI-3385 Change Testray Product Version IDs to longs instead of ints

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayProductVersion.java
@@ -47,8 +47,8 @@ public class TestrayProductVersion {
 		}
 	}
 
-	public int getID() {
-		return _jsonObject.getInt("testrayProductVersionId");
+	public long getID() {
+		return _jsonObject.getLong("testrayProductVersionId");
 	}
 
 	public String getName() {
